flask/requirements.txt
flask
mysql-connector
mysql-connector-python